refactor: Batch C — decompose the storage.rs god-file (T-STORE-SPLIT)
Split the 8,247-line storage.rs into per-entity submodules under a new
`storage/` directory. The giant `impl Storage` is divided across additional
`impl Storage` blocks (Rust attaches them to the same type regardless of file),
so public methods stay callable unchanged; only one private helper called
across the new seam (record_last_backup_at) was widened to pub(crate). storage.rs
is now 4,125 lines (-50%); all 614 backend tests pass after each extraction.

New submodules:
- storage/tests.rs        (2,467) — the inline #[cfg(test)] module, verbatim
- storage/connections.rs  (1,150) — connection records, folders, credentials
- storage/settings.rs       (508) — all get/update settings accessors

The remaining impl Storage (core lifecycle, schema, dashboard accessors) and the
free settings-validation helpers are candidates for further extraction as
follow-up; storage.rs is materially smaller and navigable now.

Verified: cargo test (614) and npm run check (lint + 187 tests + tsc) green.
